+++ src/main/java/frc/robot/commands/auto/DriveForward.java
@@ -7,7 +7,7 @@ import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.Constants.WRIST;
 import frc.robot.commands.swerve.SetSwerveNeutralMode;
-import frc.robot.commands.wrist.AutoSetWristDesiredSetpoint;
+import frc.robot.commands.wrist.AutoSetWristSetpoint;
 import frc.robot.simulation.FieldSim;
 import frc.robot.subsystems.SwerveDrive;
 import frc.robot.subsystems.Wrist;
@@ -30,7 +30,7 @@ public class DriveForward extends SequentialCommandGroup {
     addCommands(
         //        new SetSwerveOdometry(swerveDrive, trajectory.get(0).getInitialHolonomicPose(),
         // fieldSim),
-        new AutoSetWristDesiredSetpoint(wrist, WRIST.SETPOINT.STOWED.get()),
+        new AutoSetWristSetpoint(wrist, WRIST.SETPOINT.STOWED.get()),
         new PlotAutoTrajectory(fieldSim, pathName, trajectory),
         autoPath,
         new SetSwerveNeutralMode(swerveDrive, NeutralMode.Brake)
