+++ src/main/java/frc/robot/commands/auto/TwoPiece.java
@@ -14,7 +14,7 @@ import frc.robot.commands.Intake.AutoRunIntakeCone;
 import frc.robot.commands.Intake.AutoRunIntakeCube;
 import frc.robot.commands.elevator.AutoSetElevatorSetpoint;
 import frc.robot.commands.swerve.SetSwerveNeutralMode;
-import frc.robot.commands.wrist.AutoSetWristDesiredSetpoint;
+import frc.robot.commands.wrist.AutoSetWristSetpoint;
 import frc.robot.simulation.FieldSim;
 import frc.robot.subsystems.Elevator;
 import frc.robot.subsystems.Intake;
@@ -51,23 +51,23 @@ public class TwoPiece extends SequentialCommandGroup {
         new PlotAutoTrajectory(fieldSim, pathName, m_trajectory),
         new ParallelCommandGroup(
             new AutoSetElevatorSetpoint(elevator, ELEVATOR.SETPOINT.SCORE_HIGH_CONE.get()),
-            new AutoSetWristDesiredSetpoint(wrist, WRIST.SETPOINT.SCORE_HIGH_CONE.get())),
+            new AutoSetWristSetpoint(wrist, WRIST.SETPOINT.SCORE_HIGH_CONE.get())),
         new AutoRunIntakeCone(intake, -0.8, vision, swerveDrive).withTimeout(1),
         new WaitCommand(0.3),
         new ParallelCommandGroup(
             new AutoSetElevatorSetpoint(elevator, ELEVATOR.SETPOINT.STOWED.get()),
-            new AutoSetWristDesiredSetpoint(wrist, WRIST.SETPOINT.STOWED.get())),
+            new AutoSetWristSetpoint(wrist, WRIST.SETPOINT.STOWED.get())),
         autoPath,
         new AutoRunIntakeCone(intake, 0, vision, swerveDrive),
         new PlotAutoTrajectory(fieldSim, pathName, m_trajectory),
         new ParallelCommandGroup(
             new AutoSetElevatorSetpoint(elevator, ELEVATOR.SETPOINT.SCORE_HIGH_CONE.get()),
-            new AutoSetWristDesiredSetpoint(wrist, WRIST.SETPOINT.SCORE_HIGH_CONE.get())),
+            new AutoSetWristSetpoint(wrist, WRIST.SETPOINT.SCORE_HIGH_CONE.get())),
         new AutoRunIntakeCube(intake, -0.8, vision, swerveDrive).withTimeout(1),
         new WaitCommand(0.3),
         new ParallelCommandGroup(
             new AutoSetElevatorSetpoint(elevator, ELEVATOR.SETPOINT.STOWED.get()),
-            new AutoSetWristDesiredSetpoint(wrist, WRIST.SETPOINT.STOWED.get())),
+            new AutoSetWristSetpoint(wrist, WRIST.SETPOINT.STOWED.get())),
         // new AutoRunIntakeCone(intake, 0.2, vision, swerveDrive)),
 
         // new AutoRunIntakeCone(intake, 0.2, vision, swerveDrive).withTimeout(0.5),
@@ -95,7 +95,7 @@ public class TwoPiece extends SequentialCommandGroup {
         // new WaitCommand(0.5),
         new ParallelCommandGroup(
             new AutoSetElevatorSetpoint(elevator, ELEVATOR.SETPOINT.STOWED.get()),
-            new AutoSetWristDesiredSetpoint(wrist, WRIST.SETPOINT.STOWED.get())),
+            new AutoSetWristSetpoint(wrist, WRIST.SETPOINT.STOWED.get())),
         new SetSwerveNeutralMode(swerveDrive, NeutralMode.Brake)
             .andThen(() -> swerveDrive.drive(0, 0, 0, false, false)));
   }
