+++ src/main/java/frc/robot/commands/wrist/SetWristSetpoint.java
@@ -4,18 +4,26 @@
 
 package frc.robot.commands.wrist;
 
+import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants.WRIST;
 import frc.robot.subsystems.Wrist;
+import java.util.function.DoubleSupplier;
 
-public class AutoSetWristDesiredSetpoint extends CommandBase {
+public class SetWristSetpoint extends CommandBase {
   private final Wrist m_wrist;
   private double m_setpoint;
+  private final DoubleSupplier m_input;
+
+  public SetWristSetpoint(Wrist wrist, double setpoint) {
+    this(wrist, setpoint, () -> 0);
+  }
 
   /** Creates a new RunWrist. */
-  public AutoSetWristDesiredSetpoint(Wrist wrist, double setpoint) {
+  public SetWristSetpoint(Wrist wrist, double setpoint, DoubleSupplier input) {
     m_wrist = wrist;
     m_setpoint = setpoint;
+    m_input = input;
 
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(wrist);
@@ -24,25 +32,32 @@ public class AutoSetWristDesiredSetpoint extends CommandBase {
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-
+    // m_wrist.setControlState(WRIST.STATE.USER_SETPOINT);
+    // m_wrist.setDesiredPositionRadians(m_setpoint);
     m_wrist.setUserSetpoint(true);
-    m_wrist.setControlState(WRIST.STATE.CLOSED_LOOP);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
+    m_wrist.setControlState(WRIST.STATE.CLOSED_LOOP);
     m_wrist.setDesiredPositionRadians(m_setpoint);
+
+    double joystickDeadbandOutput = MathUtil.applyDeadband(m_input.getAsDouble(), 0.1);
+    m_wrist.setUserInput(-joystickDeadbandOutput);
   }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    m_wrist.setUserSetpoint(false);
+    //    m_wrist.setControlState(WRIST.STATE.CLOSED_LOOP);
+    //    m_wrist.setDesiredPositionRadians(WRIST.SETPOINT.STOWED.get());
+  }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    // 1 degree = 0.017453293 radians
-    return (Math.abs(m_wrist.getPositionRadians() - m_setpoint) < 0.017453293);
+    return false;
   }
 }
